refactor: 패키지 구조 변경 및 reformat code 실행
