+++ tomcat/src/main/java/org/apache/catalina/Manager.java
@@ -29,7 +29,8 @@ public interface Manager {
     void add(HttpSession session);
 
     /**
-     * Return the active Session, associated with this Manager, with the specified session id (if any); otherwise return
+     * Return the active Session, associated with this Manager, with the specified session id (if any); otherwise
+     * return
      * <code>null</code>.
      *
      * @param id The session id for the session to be returned
