+++ tomcat/src/main/java/org/apache/coyote/http11/handler/LoginUriHandler.java
@@ -1,10 +1,8 @@
-package org.apache.coyote.http11.urihandler;
+package org.apache.coyote.http11.handler;
 
-import java.io.IOException;
 import java.util.regex.Pattern;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.coyote.http11.HandlerResponse;
 import org.apache.coyote.http11.httpmessage.request.HttpRequest;
 import org.apache.coyote.http11.httpmessage.response.HttpStatus;
 import org.slf4j.Logger;
