+++ tomcat/src/main/java/org/apache/coyote/http11/httpMessage/request/Http11Version.java
@@ -4,6 +4,7 @@ public enum Http11Version {
 
     HTTP_11_VERSION("HTTP/1.1");
     private final String version;
+
     Http11Version(String version) {
         this.version = version;
     }
